refactor: centralize session token verification
